Move calendar files
