+++ src/main/java/com/google/sps/servlets/CalendarServlet.java
@@ -1,8 +1,22 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package com.google.sps.servlets;
 
 import com.google.cloud.datastore.*;
 import com.google.gson.Gson;
-import com.google.sps.models.CalendarUpdate;
+import com.google.sps.data.CalendarUpdate;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
