+++ src/main/webapp/calendar.html
@@ -5,13 +5,13 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Calendar</title>
-    <link rel="stylesheet" href="./index.css" />
+    <link rel="stylesheet" href="./css/calendar.css" />
   </head>
   <body>
     <div class="year"></div>
     <button id="save-btn">Save</button>
   </body>
-  <script src="./weekdays.js"></script>
-  <script src="./renderYear.js"></script>
-  <script src="./app.js"></script>
+  <script src="./js/calendar/weekdays.js"></script>
+  <script src="./js/calendar/renderYear.js"></script>
+  <script src="./js/calendar/app.js"></script>
 </html>
